Prove multithreading (web workers) for gene cache
