+++ src/js/init/caches/cache-lib.js
@@ -1,0 +1,24 @@
+import version from '../../version';
+
+export async function cacheFetch(url) {
+
+  const cache = await caches.open(`ideogram-${version}`);
+
+  const decompressedUrl = url.replace('.gz', '');
+  const response = await cache.match(decompressedUrl);
+  if (typeof response === 'undefined') {
+    // If cache miss, then fetch and add response to cache
+    // await Ideogram.cache.add(url);
+    const rawResponse = await fetch(url);
+    const blob = await rawResponse.blob();
+    const uint8Array = new Uint8Array(await blob.arrayBuffer());
+    const data = strFromU8(decompressSync(uint8Array));
+    const decompressedResponse = new Response(
+      new Blob([data], {type: 'text/tab-separated-values'}),
+      rawResponse.init
+    );
+    await cache.put(decompressedUrl, decompressedResponse);
+    return await cache.match(decompressedUrl);
+  }
+  return await cache.match(decompressedUrl);
+}
